% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_empty_example.R
\name{add_examples_to_function_docs}
\alias{add_examples_to_function_docs}
\title{Add Examples Section to Function Documentation}
\usage{
add_examples_to_function_docs(
  existing_docs,
  function_name,
  args,
  example_template
)
}
\arguments{
\item{existing_docs}{Character. Current documentation}

\item{function_name}{Character. Name of the function}

\item{args}{List. Function arguments}

\item{example_template}{Character. Template for examples}
}
\value{
Character. Updated documentation
}
\description{
Add Examples Section to Function Documentation
}
