% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace_functions_simple.r
\name{build_function_mappings}
\alias{build_function_mappings}
\title{Build Function to Package Mappings from DESCRIPTION}
\usage{
build_function_mappings(package_path, verbose = TRUE)
}
\arguments{
\item{package_path}{Character. Package root path}

\item{verbose}{Logical. Print messages}
}
\value{
Named list mapping function names to package names
}
\description{
Build Function to Package Mappings from DESCRIPTION
}
