% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_analysis.r
\name{build_parameter_history}
\alias{build_parameter_history}
\title{Build Parameter History from Existing Documentation}
\usage{
build_parameter_history(docs_data)
}
\arguments{
\item{docs_data}{Data.frame with documentation information}
}
\value{
Named list of parameter suggestions
}
\description{
Build Parameter History from Existing Documentation
}
