% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_dependency_analysis.r
\name{check_additional_dependencies}
\alias{check_additional_dependencies}
\title{Check for Additional Hidden Dependencies}
\usage{
check_additional_dependencies(package_path)
}
\arguments{
\item{package_path}{Character. Package root}
}
\value{
List. Additional dependencies found
}
\description{
Check for Additional Hidden Dependencies
}
