% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_functions.r
\name{check_export_status}
\alias{check_export_status}
\title{Check if Package Has Exported Functions}
\usage{
check_export_status(package_path = ".")
}
\arguments{
\item{package_path}{Character. Path to package root}
}
\value{
List with export information
}
\description{
Quick check to see if the package has any exported functions.
}
