% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_dependency_analysis.r
\name{create_usage_summary}
\alias{create_usage_summary}
\title{Create Usage Summary by Package}
\usage{
create_usage_summary(function_usage)
}
\arguments{
\item{function_usage}{Data.frame. Function usage data}
}
\value{
List. Summary by package
}
\description{
Create Usage Summary by Package
}
