% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace_utils.R
\name{detect_package_root_standalone}
\alias{detect_package_root_standalone}
\title{Detect Package Root Directory}
\usage{
detect_package_root_standalone(initial_path, verbose = TRUE)
}
\arguments{
\item{initial_path}{Character. Initial path to check}

\item{verbose}{Logical. Print messages}
}
\description{
Detect Package Root Directory
}
