% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_global_variables.R
\name{determine_correct_file_path}
\alias{determine_correct_file_path}
\title{Determine Correct File Path for check_global.R}
\usage{
determine_correct_file_path(file_path)
}
\arguments{
\item{file_path}{Character. Initial file path}
}
\value{
Character. Corrected file path
}
\description{
Determine Correct File Path for check_global.R
}
