% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_empty_example.R
\name{extract_empty_examples_functions}
\alias{extract_empty_examples_functions}
\title{Extract Function Names from Empty Examples Check Output}
\usage{
extract_empty_examples_functions(check_output)
}
\arguments{
\item{check_output}{Character. Raw check output text}
}
\value{
Character vector of function names
}
\description{
Extract Function Names from Empty Examples Check Output
}
