% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_global_variables.R
\name{extract_existing_global_variables}
\alias{extract_existing_global_variables}
\title{Extract Existing Global Variables from check_global.R}
\usage{
extract_existing_global_variables(file_path)

extract_existing_global_variables(file_path)
}
\arguments{
\item{file_path}{Character. Path to existing file}
}
\value{
Character vector of existing variable names

Character vector of existing variable names
}
\description{
Extract Existing Global Variables from check_global.R

Extract Existing Global Variables from check_global.R
}
