% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_analysis.r
\name{extract_function_args_simple}
\alias{extract_function_args_simple}
\title{Extract Function Arguments (Simple Version)}
\usage{
extract_function_args_simple(file_text, func_start_pos)
}
\arguments{
\item{file_text}{Character. Full file content}

\item{func_start_pos}{Integer. Position where function definition starts}
}
\value{
Character. Arguments string
}
\description{
Extract Function Arguments (Simple Version)
}
