% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_dependency_analysis.r
\name{extract_functions_after_library}
\alias{extract_functions_after_library}
\title{Extract Functions After Library/Require Calls}
\usage{
extract_functions_after_library(type_usage, package_name)
}
\arguments{
\item{type_usage}{Data.frame. Usage data}

\item{package_name}{Character. Package name}
}
\value{
Character vector. Function names
}
\description{
Extract Functions After Library/Require Calls
}
