% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_global_variables.R
\name{extract_global_variables}
\alias{extract_global_variables}
\title{Extract Global Variable Names from R CMD check Output}
\usage{
extract_global_variables(check_output)

extract_global_variables(check_output)
}
\arguments{
\item{check_output}{Character. Raw check output text}
}
\value{
Character vector of variable names

Character vector of variable names
}
\description{
Extract Global Variable Names from R CMD check Output

Extract Global Variable Names from R CMD check Output
}
