% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_dependency_analysis.r
\name{extract_namespace_functions}
\alias{extract_namespace_functions}
\title{Extract Function Names from Namespace Calls}
\usage{
extract_namespace_functions(type_usage, package_name)
}
\arguments{
\item{type_usage}{Data.frame. Usage data for namespace type}

\item{package_name}{Character. Package name}
}
\value{
Character vector. Function names
}
\description{
Extract Function Names from Namespace Calls
}
