% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_dependency_analysis.r
\name{extract_namespace_functions_from_file}
\alias{extract_namespace_functions_from_file}
\title{Extract Namespace Functions from Single File}
\usage{
extract_namespace_functions_from_file(file_path, package_name)
}
\arguments{
\item{file_path}{Character. Path to file}

\item{package_name}{Character. Package name}
}
\value{
Character vector. Function names
}
\description{
Extract Namespace Functions from Single File
}
