% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_dependency_analysis.r
\name{extract_roxygen_functions}
\alias{extract_roxygen_functions}
\title{Extract Functions from Roxygen @importFrom}
\usage{
extract_roxygen_functions(type_usage, package_name)
}
\arguments{
\item{type_usage}{Data.frame. Usage data}

\item{package_name}{Character. Package name}
}
\value{
Character vector. Function names
}
\description{
Extract Functions from Roxygen @importFrom
}
