% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_dependency_analysis.r
\name{extract_specific_functions_from_code}
\alias{extract_specific_functions_from_code}
\title{Extract Specific Functions from Code for a Package}
\usage{
extract_specific_functions_from_code(pkg_usage, package_path, package_name)
}
\arguments{
\item{pkg_usage}{Data.frame. Usage data for specific package}

\item{package_path}{Character. Package path}

\item{package_name}{Character. Package name}
}
\value{
List. Detailed function extraction
}
\description{
Extract Specific Functions from Code for a Package
}
