% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_dependency_analysis.r
\name{extract_suspected_functions_from_file}
\alias{extract_suspected_functions_from_file}
\title{Extract Suspected Functions from File}
\usage{
extract_suspected_functions_from_file(file_path, package_name)
}
\arguments{
\item{file_path}{Character. File path}

\item{package_name}{Character. Package name}
}
\value{
Character vector. Function names
}
\description{
Extract Suspected Functions from File
}
