% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_functions.r
\name{filter_exported_functions}
\alias{filter_exported_functions}
\title{Filter Functions to Only Include Exported/Export-Intended Functions}
\usage{
filter_exported_functions(function_names, func_data, docs_data)
}
\arguments{
\item{function_names}{Character vector. Function names to filter}

\item{func_data}{Data.frame. Function data subset}

\item{docs_data}{Data.frame. Documentation data subset}
}
\value{
Character vector of export-intended function names
}
\description{
Identifies functions that either already have @export tags or appear to be
intended for export based on the context.
}
