% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_analysis.r
\name{find_docs_start_simple}
\alias{find_docs_start_simple}
\title{Find Documentation Start (Simple Version)}
\usage{
find_docs_start_simple(file_text, func_start_pos)
}
\arguments{
\item{file_text}{Character. Full file content}

\item{func_start_pos}{Integer. Position where function starts}
}
\value{
List with start_pos and has_docs
}
\description{
Find Documentation Start (Simple Version)
}
