% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_dependency_analysis.r
\name{find_function_at_position}
\alias{find_function_at_position}
\title{Find Function at Specific Position in File}
\usage{
find_function_at_position(file_text, position, file_functions)
}
\arguments{
\item{file_text}{Character. Full file content}

\item{position}{Integer. Character position}

\item{file_functions}{Data.frame. Functions in this file}
}
\value{
Character. Function name or NA
}
\description{
Find Function at Specific Position in File
}
