% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_dependency_analysis.r
\name{find_functions_using_package_patterns}
\alias{find_functions_using_package_patterns}
\title{Find Functions Using Package Patterns}
\usage{
find_functions_using_package_patterns(file_text, package, file_functions)
}
\arguments{
\item{file_text}{Character. File content}

\item{package}{Character. Package name}

\item{file_functions}{Data.frame. Functions in file}
}
\value{
Character vector. Function names
}
\description{
Find Functions Using Package Patterns
}
