% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace_functions_simple.r
\name{find_namespace_opportunities}
\alias{find_namespace_opportunities}
\title{Find Namespace Conversion Opportunities in File Content}
\usage{
find_namespace_opportunities(content, mappings)
}
\arguments{
\item{content}{Character vector. File lines}

\item{mappings}{Named list. Function to package mappings}
}
\value{
List of opportunities
}
\description{
Find Namespace Conversion Opportunities in File Content
}
