% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibecheck_wrapper.r
\name{fix_documentation_interactive}
\alias{fix_documentation_interactive}
\title{Interactive Documentation Fix}
\usage{
fix_documentation_interactive(pkg_info)
}
\arguments{
\item{pkg_info}{List. Package analysis results}
}
\description{
Interactive Documentation Fix
}
