% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_empty_example.R
\name{fix_empty_examples_interactive}
\alias{fix_empty_examples_interactive}
\title{Interactive Function to Fix Empty Examples}
\usage{
fix_empty_examples_interactive(
  example_template = "x <- FUNCNAME(data)",
  package_path = "."
)
}
\arguments{
\item{example_template}{Character. Template for examples}

\item{package_path}{Character. Path to package}
}
\value{
Result from fix_empty_examples()
}
\description{
Interactive Function to Fix Empty Examples
}
