% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibecheck_wrapper.r
\name{fix_missing_packages_interactive}
\alias{fix_missing_packages_interactive}
\title{Interactive Missing Packages Fix}
\usage{
fix_missing_packages_interactive(pkg_info)
}
\arguments{
\item{pkg_info}{List. Package analysis results}
}
\description{
Interactive Missing Packages Fix
}
