% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibecheck_wrapper.r
\name{fix_namespace_interactive}
\alias{fix_namespace_interactive}
\title{Interactive Namespace Fix}
\usage{
fix_namespace_interactive(namespace_analysis)
}
\arguments{
\item{namespace_analysis}{List. Namespace analysis results}
}
\description{
Interactive Namespace Fix
}
