% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_functions.r
\name{generate_custom_example}
\alias{generate_custom_example}
\title{Generate Custom Example from Template}
\usage{
generate_custom_example(function_name, example_template)
}
\arguments{
\item{function_name}{Character. Function name}

\item{example_template}{Character. Template with \link{FUNCNAME} placeholder}
}
\value{
Character. Custom example with function name substituted
}
\description{
Replaces \link{FUNCNAME} placeholder in template with actual function name.
}
\examples{
\dontrun{
template <- "dhsData <- getDHSdata(country = \"Zambia\", indicator = \"[FUNCNAME]\", year = 2018)"
example <- generate_custom_example("get_mortality", template)
# Returns: "dhsData <- getDHSdata(country = \"Zambia\", indicator = \"get_mortality\", year = 2018)"
}

}
