% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_functions.r
\name{generate_function_example}
\alias{generate_function_example}
\title{Generate Smart Function Example}
\usage{
generate_function_example(function_name, args)
}
\arguments{
\item{function_name}{Character. Function name}

\item{args}{Named list. Function arguments}
}
\value{
Character. Example usage
}
\description{
Generate Smart Function Example
}
