% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_global_variables.R
\name{generate_global_variables_file}
\alias{generate_global_variables_file}
\title{Generate Content for check_global.R File}
\usage{
generate_global_variables_file(variables)

generate_global_variables_file(variables)
}
\arguments{
\item{variables}{Character vector. All variable names to include}
}
\value{
Character vector. Lines of file content

Character vector. Lines of file content
}
\description{
Generate Content for check_global.R File

Generate Content for check_global.R File
}
