% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace_functions_simple.r
\name{generate_namespace_report}
\alias{generate_namespace_report}
\title{Generate Namespace Conversion Report}
\usage{
generate_namespace_report(analysis_result)
}
\arguments{
\item{analysis_result}{List. Result from analyze_namespace_usage()}
}
\value{
Character. Formatted report
}
\description{
Generate Namespace Conversion Report
}
