% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_functions.r
\name{get_exported_functions}
\alias{get_exported_functions}
\title{Get Exported Functions from NAMESPACE File}
\usage{
get_exported_functions(package_path)
}
\arguments{
\item{package_path}{Character. Path to package root}
}
\value{
Character vector of exported function names
}
\description{
Reads the NAMESPACE file and extracts all exported function names.
}
