% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_dependency_analysis.r
\name{get_package_name_from_description}
\alias{get_package_name_from_description}
\title{Get Package Name from DESCRIPTION File}
\usage{
get_package_name_from_description(package_path)
}
\arguments{
\item{package_path}{Character. Path to package root}
}
\value{
Character. Package name or NULL if not found
}
\description{
Get Package Name from DESCRIPTION File
}
