% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibecheck_wrapper.r
\name{interactive_fix_menu}
\alias{interactive_fix_menu}
\title{Interactive Fix Menu}
\usage{
interactive_fix_menu(pkg_info, namespace_analysis = NULL)
}
\arguments{
\item{pkg_info}{List. Package analysis results}

\item{namespace_analysis}{List. Namespace analysis results}
}
\description{
Interactive Fix Menu
}
