% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_functions.r
\name{is_function_likely_exported}
\alias{is_function_likely_exported}
\title{Check if Function is Likely Intended for Export}
\usage{
is_function_likely_exported(func_name, func_row)
}
\arguments{
\item{func_name}{Character. Function name}

\item{func_row}{Single row data.frame with function information}
}
\value{
Logical. TRUE if function appears intended for export
}
\description{
Uses naming conventions and other heuristics to determine if an undocumented
function is likely intended to be part of the public API.
}
