% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_analysis.r
\name{is_package_root}
\alias{is_package_root}
\title{Check if Directory is Package Root}
\usage{
is_package_root(path)
}
\arguments{
\item{path}{Character. Path to check}
}
\value{
Logical. TRUE if package root
}
\description{
Check if Directory is Package Root
}
