% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace_utils.R
\name{is_package_root_standalone}
\alias{is_package_root_standalone}
\title{Check if Directory is Package Root}
\usage{
is_package_root_standalone(path)
}
\arguments{
\item{path}{Character. Path to check}
}
\description{
Check if Directory is Package Root
}
