% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_functions.r
\name{is_raw_content_empty}
\alias{is_raw_content_empty}
\title{Check if Raw Content (without dontrun) is Empty or Just Placeholders}
\usage{
is_raw_content_empty(content)
}
\arguments{
\item{content}{Character. Clean content without comment markers}
}
\value{
Logical. TRUE if content is empty or just placeholders
}
\description{
Check if Raw Content (without dontrun) is Empty or Just Placeholders
}
