% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_empty_example.R
\name{match_functions_to_files}
\alias{match_functions_to_files}
\title{Match Function Names to R Files}
\usage{
match_functions_to_files(function_names, functions_data)
}
\arguments{
\item{function_names}{Character vector. Function names to find}

\item{functions_data}{Data frame. Function data from DocumentationHelper}
}
\value{
Named character vector. Names are function names, values are file paths
}
\description{
Match Function Names to R Files
}
