% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_dependency_analysis.r
\name{merge_additional_deps}
\alias{merge_additional_deps}
\title{Merge Additional Dependencies}
\usage{
merge_additional_deps(all_deps, additional_deps)
}
\arguments{
\item{all_deps}{List. Main dependencies}

\item{additional_deps}{List. Additional dependencies}
}
\value{
List. Merged dependencies
}
\description{
Merge Additional Dependencies
}
