% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibecheck_wrapper.r
\name{offer_interactive_fixes}
\alias{offer_interactive_fixes}
\title{Offer Interactive Fixes for Common Issues}
\usage{
offer_interactive_fixes(pkg_info, verbose = TRUE)
}
\arguments{
\item{pkg_info}{List. Package analysis results}

\item{verbose}{Logical. Print messages}
}
\description{
Offer Interactive Fixes for Common Issues
}
