% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_analysis.r
\name{parse_description_dependencies}
\alias{parse_description_dependencies}
\title{Parse DESCRIPTION File Dependencies}
\usage{
parse_description_dependencies(package_path)
}
\arguments{
\item{package_path}{Character. Package root path}
}
\value{
List with declared dependencies
}
\description{
Parse DESCRIPTION File Dependencies
}
