% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_analysis.r
\name{parse_function_docs}
\alias{parse_function_docs}
\title{Parse Function Documentation}
\usage{
parse_function_docs(functions_data)
}
\arguments{
\item{functions_data}{Data.frame from scan_r_functions}
}
\value{
Data.frame with documentation information
}
\description{
Parse Function Documentation
}
