% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibecheck_wrapper.r
\name{print_package_summary}
\alias{print_package_summary}
\title{Print Package Analysis Summary}
\usage{
print_package_summary(pkg_info, verbose = TRUE)
}
\arguments{
\item{pkg_info}{List. Result from analyze_package()}

\item{verbose}{Logical. Print detailed information}
}
\description{
Print Package Analysis Summary
}
