% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_analysis.r
\name{scan_code_dependencies}
\alias{scan_code_dependencies}
\title{Scan Code for Package Dependencies}
\usage{
scan_code_dependencies(functions_data, package_path)
}
\arguments{
\item{functions_data}{Data.frame. Function information}

\item{package_path}{Character. Package root path}
}
\value{
List with detected dependencies
}
\description{
Scan Code for Package Dependencies
}
