% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibecheck_wrapper.r
\name{show_dependency_report}
\alias{show_dependency_report}
\title{Show Dependency Report}
\usage{
show_dependency_report(pkg_info)
}
\arguments{
\item{pkg_info}{List. Package analysis results}
}
\description{
Show Dependency Report
}
