% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibecheck_wrapper.r
\name{show_namespace_report}
\alias{show_namespace_report}
\title{Show Namespace Report}
\usage{
show_namespace_report(namespace_analysis)
}
\arguments{
\item{namespace_analysis}{List. Namespace analysis results}
}
\description{
Show Namespace Report
}
