% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_functions.r
\name{suggest_parameter_description}
\alias{suggest_parameter_description}
\title{Suggest Parameter Description}
\usage{
suggest_parameter_description(param_name, param_suggestions = NULL)
}
\arguments{
\item{param_name}{Character. Parameter name}

\item{param_suggestions}{Named list. Historical parameter descriptions}
}
\value{
Character. Suggested description
}
\description{
Suggest Parameter Description
}
