% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_functions.r
\name{suggest_parameters}
\alias{suggest_parameters}
\title{Get Parameter Suggestions for a Function}
\usage{
suggest_parameters(package_data = NULL, param_name, function_name = NULL)
}
\arguments{
\item{package_data}{List. Result from analyze_package() (optional)}

\item{param_name}{Character. Parameter name to get suggestions for}

\item{function_name}{Character. Function name context (optional)}
}
\value{
Character vector of suggestions
}
\description{
Get Parameter Suggestions for a Function
}
\examples{
\dontrun{
# Get suggestions for a parameter
pkg_info <- analyze_package()
suggestions <- suggest_parameters(pkg_info, "data")

# Without package context
suggestions <- suggest_parameters(param_name = "verbose")
}

}
