% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_functions.r
\name{test_examples_only_update}
\alias{test_examples_only_update}
\title{Test Examples-Only Update}
\usage{
test_examples_only_update(function_name = "WS_WTRT_P_BLC")
}
\arguments{
\item{function_name}{Character. Function to test}
}
\description{
Test Examples-Only Update
}
