% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_analysis.r
\name{detect_package_root}
\alias{detect_package_root}
\title{Detect Package Root Directory}
\usage{
detect_package_root(path, verbose = TRUE)
}
\arguments{
\item{path}{Character. Initial path}

\item{verbose}{Logical. Print messages}
}
\value{
Character. Package root path
}
\description{
Detect Package Root Directory
}
