% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_analysis.r
\name{detect_suspected_packages}
\alias{detect_suspected_packages}
\title{Detect Suspected Package Usage from Code Patterns}
\usage{
detect_suspected_packages(code)
}
\arguments{
\item{code}{Character. Code content}
}
\value{
Character vector of suspected packages
}
\description{
Detect Suspected Package Usage from Code Patterns
}
