% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_analysis.r
\name{extract_function_docs}
\alias{extract_function_docs}
\title{Extract Function Documentation from File}
\usage{
extract_function_docs(file_path, start_pos)
}
\arguments{
\item{file_path}{Character. Path to file}

\item{start_pos}{Integer. Starting position of function}
}
\value{
Character. Documentation string
}
\description{
Extract Function Documentation from File
}
