% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_analysis.r
\name{generate_dependency_report}
\alias{generate_dependency_report}
\title{Generate Dependency Report}
\usage{
generate_dependency_report(dependency_data)
}
\arguments{
\item{dependency_data}{List. Result from analyze_package_dependencies}
}
\value{
Character. Formatted report
}
\description{
Generate Dependency Report
}
