% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_analysis.r
\name{scan_r_functions}
\alias{scan_r_functions}
\title{Scan R Functions in Package}
\usage{
scan_r_functions(package_path, verbose = TRUE)
}
\arguments{
\item{package_path}{Character. Package root path}

\item{verbose}{Logical. Print messages}
}
\value{
Data.frame with function information
}
\description{
Updated to scan both .R and .r files (case insensitive).
}
